AuthorizedSamples & Species per lake : Only allow selected lakes modification in REST APIs

+++ fishola-backend/src/main/java/fr/inrae/fishola/rest/referential/ReferentialResource.java
@@ -29,6 +29,7 @@ import com.google.common.collect.Sets;
 import fr.inrae.fishola.database.CatchsDao;
 import fr.inrae.fishola.database.ReferentialDao;
 import fr.inrae.fishola.entities.tables.pojos.AuthorizedSample;
+import fr.inrae.fishola.entities.tables.pojos.FisholaAdmin;
 import fr.inrae.fishola.entities.tables.pojos.Lake;
 import fr.inrae.fishola.entities.tables.pojos.ReleasedFishState;
 import fr.inrae.fishola.entities.tables.pojos.Species;
@@ -267,12 +268,12 @@ public class ReferentialResource extends AbstractFisholaResource {
 
     @PUT
     @Path("/species-aliases-per-lake")
-    public Response saveSpeciesAliasesPerLake(Map<UUID, Map<UUID, String>> aliases) {
-        checkIsAdmin();
+    public Response saveSpeciesAliasesPerLake(SpeciesAliasesPerLakeBean salp) {
+        FisholaAdmin fisholaAdmin = checkIsAdmin();
 
         // On transforme la map pour avoir en clé lakeId+speciesId et en valeur les alias
         Map<Pair<UUID, UUID>, String> aliasesMap = new HashMap<>();
-        for (Map.Entry<UUID, Map<UUID, String>> byLakeEntry : aliases.entrySet()) {
+        for (Map.Entry<UUID, Map<UUID, String>> byLakeEntry : salp.speciesPerLakeAliases.entrySet()) {
             Map<UUID, String> bySpeciesEntries = byLakeEntry.getValue();
             for (Map.Entry<UUID, String> entry : bySpeciesEntries.entrySet()) {
                 Pair<UUID, UUID> lakePluSpeciesId = Pair.of(byLakeEntry.getKey(), entry.getKey());
@@ -288,20 +289,23 @@ public class ReferentialResource extends AbstractFisholaResource {
 
         // On commence par mettre à jour ou supprimer les espèces par lac existantes
         for (SpeciesByLake entity : speciesByLake) {
-            Pair<UUID, UUID> lakePluSpeciesId = Pair.of(entity.getLakeId(), entity.getSpeciesId());
-            if (aliasesMap.containsKey(lakePluSpeciesId)) {
-                String newAlias = aliases.get(entity.getLakeId()).get(entity.getSpeciesId());
-                entity.setAlias(newAlias);
-                referentialDao.updateSpeciesByLake(entity);
-            } else {
-                referentialDao.deleteSpeciesByLake(entity);
+            if (salp.targetLakes.contains(entity.getLakeId())) {
+                Pair<UUID, UUID> lakePluSpeciesId = Pair.of(entity.getLakeId(), entity.getSpeciesId());
+                if (aliasesMap.containsKey(lakePluSpeciesId)) {
+                    String newAlias = salp.speciesPerLakeAliases.get(entity.getLakeId()).get(entity.getSpeciesId());
+                    entity.setAlias(newAlias);
+                    referentialDao.updateSpeciesByLake(entity);
+                } else {
+                    referentialDao.deleteSpeciesByLake(entity);
+                }
+                aliasesMap.remove(lakePluSpeciesId);
             }
-            aliasesMap.remove(lakePluSpeciesId);
         }
 
         // Puis on créé les nouvelles
         aliasesMap.entrySet()
                 .stream()
+                .filter(entry -> salp.targetLakes.contains(entry.getKey().getKey()))
                 .map(entry -> {
                     UUID lakeId = entry.getKey().getKey();
                     UUID speciesId = entry.getKey().getValue();
@@ -309,7 +313,9 @@ public class ReferentialResource extends AbstractFisholaResource {
                     SpeciesByLake result = new SpeciesByLake(lakeId, speciesId, alias);
                     return result;
                 })
-                .forEach(referentialDao::createSpeciesByLake);
+                .forEach(spl -> {
+                    referentialDao.createSpeciesByLake(spl);
+                });
 
         Response response = Response.noContent().build();
         return response;
@@ -317,16 +323,20 @@ public class ReferentialResource extends AbstractFisholaResource {
 
     @PUT
     @Path("/authorized-samples")
-    public Response saveAuthorizedSamples(List<Map<UUID, Map<UUID, Object>>> authorizationsAndMinMaxSizes) {
-        checkIsAdmin();
-
+    public Response saveAuthorizedSamples(AuthorizedSamplesModificationBean authorizedSamples) {
+        FisholaAdmin fisholaAdmin = checkIsAdmin();
+        Set<UUID> allowedAdminLakes = getAllowedAdminLakes();
+        Set<UUID> lakeScope =  authorizedSamples.targetLakes.stream()
+            .filter(l -> fisholaAdmin.getIsnationaladmin() || allowedAdminLakes.contains(l))
+            .collect(Collectors.toSet());
+        
         // On transforme la map pour avoir un Set des clé lakeId+speciesId autorisées
         Set<Pair<UUID, UUID>> authorizationsSet = new HashSet<>();
         Map<Pair<UUID, UUID>, Integer> minSizesMap = new LinkedHashMap<>();
         Map<Pair<UUID, UUID>, Integer> maxSizesMap = new LinkedHashMap<>();
-        Map<UUID, Map<UUID, Object>> authorizations = authorizationsAndMinMaxSizes.get(0);
-        Map<UUID, Map<UUID, Object>> minSizes = authorizationsAndMinMaxSizes.get(1);
-        Map<UUID, Map<UUID, Object>> maxSizes = authorizationsAndMinMaxSizes.get(2);
+        Map<UUID, Map<UUID, Object>> authorizations = authorizedSamples.authorizations;
+        Map<UUID, Map<UUID, Object>> minSizes = authorizedSamples.minSizes;
+        Map<UUID, Map<UUID, Object>> maxSizes = authorizedSamples.maxSizes;
         for (Map.Entry<UUID, Map<UUID, Object>> byLakeEntry : authorizations.entrySet()) {
             Map<UUID, Object> bySpeciesEntries = byLakeEntry.getValue();
             for (Map.Entry<UUID, Object> entry : bySpeciesEntries.entrySet()) {
@@ -350,28 +360,31 @@ public class ReferentialResource extends AbstractFisholaResource {
         // On commence par supprimer les autorisations en trop
         for (AuthorizedSample entity : existingAuthorizations) {
             Pair<UUID, UUID> lakePluSpeciesId = Pair.of(entity.getLakeId(), entity.getSpeciesId());
-            if (!authorizationsSet.contains(lakePluSpeciesId)) {
-                referentialDao.deleteAuthorizedSample(entity);
-            } else {
-                // On met à jour uniquement la taille si l'autorisation existe dejà
-                Integer minSize = 0;
-                if (minSizesMap.get(lakePluSpeciesId) != null) {
-                    minSize = minSizesMap.get(lakePluSpeciesId);
+            if (lakeScope.contains(entity.getLakeId())) {
+                if (!authorizationsSet.contains(lakePluSpeciesId)) {
+                    referentialDao.deleteAuthorizedSample(entity);
+                } else {
+                    // On met à jour uniquement la taille si l'autorisation existe dejà
+                    Integer minSize = 0;
+                    if (minSizesMap.get(lakePluSpeciesId) != null) {
+                        minSize = minSizesMap.get(lakePluSpeciesId);
+                    }
+                    Integer maxSize = 0;
+                    if (maxSizesMap.get(lakePluSpeciesId) != null) {
+                        maxSize = maxSizesMap.get(lakePluSpeciesId);
+                    }
+                    entity.setMinSize(minSize);
+                    entity.setMaxSize(maxSize);
+                    referentialDao.updateAuthorizeSample(entity);
                 }
-                Integer maxSize = 0;
-                if (maxSizesMap.get(lakePluSpeciesId) != null) {
-                    maxSize = maxSizesMap.get(lakePluSpeciesId);
-                }
-                entity.setMinSize(minSize);
-                entity.setMaxSize(maxSize);
-                referentialDao.updateAuthorizeSample(entity);
+                authorizationsSet.remove(lakePluSpeciesId);
             }
-            authorizationsSet.remove(lakePluSpeciesId);
         }
 
         // Puis on créé les nouvelles
         authorizationsSet
             .stream()
+            .filter(entry -> lakeScope.contains(entry.getKey()))
             .map(entry -> {
                 Integer minSize = 0;
                 Integer maxSize = 0;
